backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
elevenlabs
pillow
sqlalchemy
pydantic
pydantic-settings
python-multipart